make credentials faster
The google/auth package provides a general-purpose caching
with FetchAuthTokenCache. The implementation is aimed to be compatible
with PSR-6.

This commit introduces a faster cache, usable in conjuction with
FetchAuthTokenCache.
It queries for the last received token and reuses it if the token has
not expired.
In this way, we avoid sanitizing cache keys and performing table
lookups.

In my testing, this increases our throughput by ~10%.