test: Fix retry test that assumed floating-point timeouts
RetryMiddlewareTest::testRetryLogicalTimeout depended on the
passage of real-world time to assert that the the timeout
decreases with each attempt.

Previously this could reliably assert that each timeout
was < the previous because they were floating point values. So
even a few microseconds to execute each loop of the test was
enough for the value to change.

However now that the timeout is rounded down to an int, the
test was moving too fast to see a decrease. Added a usleep
to ensure that each loop takes at least 1ms, so the timeout
is lower each time.